reimagination of battlesystem & implementation
